feat(resultsTableApiConnect): rewrote Home.stories.jsx
